Removed all remaining demag logic.
Demag logic was there in case there was a desync caused by a fast acceleration.
If the program were doing the only task of monitoring commutation there won't be any problem, but as the program is doing other tasks there could exist the possibility that it can miss zero crossings.
Those zero crossings misses were the origin of demag terminology and demag metric.
This commit removes the last bytes of that commutation strategy. From now commutation is simpler.